Implemented functionality to read and parse data returned by SQLite
SQLite in python returns each table as a list of tuples. Here tuples are the rows of each table and each element in a tuple is a value in a database row.
To pass it to JS, the python lists (tables) are parsed into JSON. The JSON string is then fetched by JS and is parsed as a readable object.